🎨:: Renamed Folder Images to Assets, Additional File Structure Improvements, Tags Re-Format
