+++ src/Assets/svgs/Search.jsx
