+++ src/Assets/svgs/index.jsx
@@ -1,0 +1,3 @@
+export { default as BlueCircleSVG } from './BlueCircle.jsx';
+export { default as GrayCircleSVG } from './GrayCircle.jsx';
+export { default as SearchSVG } from './Search.jsx';
